requirements/dev.txt
# Local development dependencies go here
-r base.txt

coverage
flake8
mock
pydocstyle
pytest-pep8
pytest-cov
Sphinx
pyserial
crcmod
statsd
